requirements.txt
langchain==0.2.16
langchain_community==0.2.16
langchain_core==0.2.38
langchain_google_genai==1.0.10
langchain_openai==0.1.23
python-dotenv==1.0.1
streamlit==1.37.0
